+++ src/activities/settings/StatusBarSettingsActivity.h
@@ -9,8 +9,8 @@
 // Reader status bar configuration activity
 class StatusBarSettingsActivity final : public Activity {
  public:
-  explicit StatusBarSettingsActivity(GfxRenderer& renderer, MappedInputManager& mappedInput)
-      : Activity("StatusBarSettings", renderer, mappedInput) {}
+  explicit StatusBarSettingsActivity(GfxRenderer& renderer, MappedInputManager& mappedInput, bool readerContext = false)
+      : Activity("StatusBarSettings", renderer, mappedInput), readerContext(readerContext) {}
 
   void onEnter() override;
   void onExit() override;
@@ -22,6 +22,7 @@ class StatusBarSettingsActivity final : public Activity {
 
   int selectedIndex = 0;
   int visibleItemCount = 0;
+  bool readerContext = false;
 
   void handleSelection();
 };